--- a/WPF/SeeShells/SeeShells/UI/Pages/TimelinePage.xaml.cs
+++ b/WPF/SeeShells/SeeShells/UI/Pages/TimelinePage.xaml.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using System.Windows.Shapes;
 using Xceed.Wpf.Toolkit;
 using Xceed.Wpf.Toolkit.Primitives;
 
@@ -217,6 +218,40 @@ namespace SeeShells.UI.Pages
             DateTime beginDate = DateTimeRoundDown(nodesCluster[0].aEvent.EventTime, maxRealTimeSpan);
             DateTime endDate = beginDate.AddMinutes(1);
 
+            TimelinePanel timelinePanel = MakeTimelinePanel(beginDate, endDate);
+
+            List<StackedNodes> stackedNodesList = GetStackedNodes(nodesCluster);
+            // Add all nodes that stack onto a timeline
+            foreach (StackedNodes stackedNode in stackedNodesList)
+            {
+                TimelinePanel.SetDate(stackedNode, stackedNode.events[0].EventTime);
+                stackedNode.Content = stackedNode.events.Count.ToString();
+                timelinePanel.Children.Add(stackedNode);
+                ConnectNodeToTimeline(timelinePanel, stackedNode.events[0].EventTime);
+            }
+            // Add all other nodes onto a timeline
+            foreach (Node.Node node in nodesCluster)
+            {
+                TimelinePanel.SetDate(node, node.aEvent.EventTime);
+                timelinePanel.Children.Add(node);
+                ConnectNodeToTimeline(timelinePanel, node.aEvent.EventTime);
+            }
+
+            Timelines.Children.Add(timelinePanel);
+            Line separationLine = MakeTimelineSeparatingLine();
+            Timelines.Children.Add(separationLine);
+            AddTicks(beginDate, endDate);
+            AddTimeStamp(beginDate, endDate);
+        }
+
+        /// <summary>
+        /// Creates a TimelinePanel
+        /// </summary>
+        /// <param name="beginDate">begin date of timeline</param>
+        /// <param name="endDate">end date of timeline</param>
+        /// <returns>TimelinePanel that can space graphical objects according to time</returns>
+        private TimelinePanel MakeTimelinePanel(DateTime beginDate, DateTime endDate)
+        {
             TimelinePanel timelinePanel = new TimelinePanel
             {
                 UnitTimeSpan = new TimeSpan(0, 0, 0, 1),
@@ -226,23 +261,7 @@ namespace SeeShells.UI.Pages
                 KeepOriginalOrderForOverlap = true
             };
 
-            List<StackedNodes> stackedNodesList = GetStackedNodes(nodesCluster);
-            // Add all nodes that stack onto a timeline
-            foreach (StackedNodes stackedNode in stackedNodesList)
-            {
-                TimelinePanel.SetDate(stackedNode, stackedNode.events[0].EventTime);
-                stackedNode.Content = stackedNode.events.Count.ToString();
-                timelinePanel.Children.Add(stackedNode);
-            }
-            // Add all other nodes onto a timeline
-            foreach (Node.Node node in nodesCluster)
-            {
-                TimelinePanel.SetDate(node, node.aEvent.EventTime);
-                timelinePanel.Children.Add(node);
-            }
-
-            Timelines.Children.Add(timelinePanel);
-            AddTextBlockTimeStamp(beginDate, endDate);
+            return timelinePanel;
         }
 
         /// <summary>
@@ -290,19 +309,107 @@ namespace SeeShells.UI.Pages
         }
 
         /// <summary>
-        /// Adds a timestamp for timeline.
+        /// Draws a line to connect a node to a timeline
         /// </summary>
-        /// <param name="beginDate">the begin date of the time interval</param>
-        /// <param name="endDate">the end date of the time interval</param>
-        private void AddTextBlockTimeStamp(DateTime beginDate, DateTime endDate)
+        /// <param name="timelinePanel">timeline panel to hold and position connection lines</param>
+        /// <param name="eventTime">time used as position for where to draw a connecting line</param>
+        private void ConnectNodeToTimeline(TimelinePanel timelinePanel, DateTime eventTime)
         {
-            TextBlock textBlock = new TextBlock();
-            textBlock.Text = beginDate.ToString() + " - " + endDate.ToString();
-            textBlock.Height = 20;
-            textBlock.Width = endDate.Subtract(beginDate).TotalSeconds * App.nodeCollection.nodeList[0].Width;
-            textBlock.Background = Brushes.LightSteelBlue;
+            Line connectorLine = new Line();
+            connectorLine.Stroke = Brushes.LightSteelBlue;
+            connectorLine.X1 = 5;
+            connectorLine.X2 = 5;
+            connectorLine.Y1 = 0;
+            connectorLine.Y2 = 13;
+            connectorLine.StrokeThickness = 1;
 
-            TimeStamps.Children.Add(textBlock);
+            TimelinePanel.SetDate(connectorLine, eventTime);
+            timelinePanel.Children.Add(connectorLine);
+        }
+
+        /// <summary>
+        /// Draws ticks below the nodes to represent the seconds of each timeline interval
+        /// </summary>
+        /// <param name="beginDate">begin date of a timeline period</param>
+        /// <param name="endDate"> end date of a timeline period</param>
+        private void AddTicks(DateTime beginDate, DateTime endDate)
+        {
+            TimelinePanel timelinePanel = MakeTimelinePanel(beginDate, endDate);
+
+            AddTicksBar(beginDate, endDate);
+            int timePeriod = (int)endDate.Subtract(beginDate).TotalSeconds;
+            for (int i = 0; i < timePeriod; i++)
+            {
+                Line tick = new Line();
+                tick.Stroke = Brushes.Black;
+                tick.X1 = 5;
+                tick.X2 = 5;
+                tick.Y1 = 0;
+                tick.Y2 = 20;
+                tick.StrokeThickness = 1;
+
+                TimelinePanel.SetDate(tick, beginDate);
+                timelinePanel.Children.Add(tick);
+                beginDate = beginDate.AddSeconds(1);
+            }
+            Ticks.Children.Add(timelinePanel);
+
+            Line separationLine = MakeTimelineSeparatingLine();
+            separationLine.Visibility = Visibility.Hidden; // Hidden since this line is added only for proper spacing
+            Ticks.Children.Add(separationLine);
+        }
+
+        /// <summary>
+        /// Adds a rectangular bar behind the timeline ticks.
+        /// </summary>
+        /// <param name="beginDate">begin date of the time interval used to decide the size of the bar</param>
+        /// <param name="endDate">end date of the time interval used to decide the size of the bar</param>
+        private void AddTicksBar(DateTime beginDate, DateTime endDate)
+        {
+            Rectangle bar = new Rectangle();
+            bar.Height = 20;
+            bar.Width = (endDate.Subtract(beginDate).TotalSeconds * App.nodeCollection.nodeList[0].Width) + 12; // The added integer is to compensate for the margins and thickness of the line that separates the timelines. 
+            bar.Fill = Brushes.LightSteelBlue;
+            bar.Margin = new Thickness(0, 0, -1, 0);
+
+            TicksBar.Children.Add(bar);
+        }
+
+        /// <summary>
+        /// Adds time stamp that tells the time period of a timeline
+        /// </summary>
+        /// <param name="beginDate">the begin date of the timeline</param>
+        /// <param name="endDate">the end date of the timeline</param>
+        private void AddTimeStamp(DateTime beginDate, DateTime endDate)
+        {
+            TextBlock timeStamp = new TextBlock();
+            timeStamp.Text = beginDate.ToString() + " - " + endDate.ToString();
+            timeStamp.Foreground = Brushes.White;
+            timeStamp.Height = 20;
+            timeStamp.Width = (endDate.Subtract(beginDate).TotalSeconds * App.nodeCollection.nodeList[0].Width) + 12;
+            timeStamp.Margin = new Thickness(0, 0, -1, 0);
+
+            TimeStamps.Children.Add(timeStamp);
+        }
+
+        /// <summary>
+        /// Creates a line to be used as separation between timelines
+        /// </summary>
+        /// <returns>a line to visually separate timelines</returns>
+        private Line MakeTimelineSeparatingLine()
+        {
+            Line separatingLine = new Line();
+            separatingLine.Stroke = Brushes.LightSteelBlue;
+            separatingLine.X1 = 0;
+            separatingLine.X2 = 0;
+            separatingLine.Y1 = 43;
+            separatingLine.Y2 = 150;
+            separatingLine.StrokeThickness = 2;
+            separatingLine.HorizontalAlignment = HorizontalAlignment.Left;
+            separatingLine.VerticalAlignment = VerticalAlignment.Center;
+            separatingLine.Margin = new Thickness(5, 0, 5, 0);
+
+            return separatingLine;
         }
 
         /// <summary>
@@ -318,15 +425,21 @@ namespace SeeShells.UI.Pages
         }
 
         /// <summary>
-        /// Clears the children of all timeline related UI objects and builds timeline.
+        /// Clears all children of UI objects and rebuilds the timeline.
         /// </summary>
         public void RebuildTimeline()
         {
-            foreach(TimelinePanel timeline in Timelines.Children)
+            foreach (Object child in Timelines.Children)
             {
-                timeline.Children.Clear();
+                if(child is TimelinePanel) // Only TimelinePanel objects since Timelines also contains separating lines
+                {
+                    TimelinePanel timeline = (TimelinePanel)child;
+                    timeline.Children.Clear();
+                }
             }
             Timelines.Children.Clear();
+            Ticks.Children.Clear();
+            TicksBar.Children.Clear();
             TimeStamps.Children.Clear();
             BuildTimeline();
         }