1. Adds ticks to graph out each second of the timeline. (the ticks are placed in a TimelinePanel which handles their spacing)
2. Adds separation lines to designate the end of one timeline and start of another.
3. Adds a timeline bar to represent the continuous timeline.
4. Adds lines to connect each node to the timeline bar in order to pinpoint their location in relation to the ticks on the bar.

Note:
Due to a significant amount of objects being added to the UI, the program has a slight slowdown. It takes about extra 3 to 4 seconds to parse and 0 to 4 seconds to redraw after filters are applied depending on how large the filtered amount of events is. However, I am running it on pretty fast hardware, thus the slowdown might be more severe on lower end machines. This will be something to test and possibly optimize in the future.

![image](https://user-images.githubusercontent.com/42561260/76688208-2c276080-6601-11ea-83c8-026bf260a6c7.png)

Resolves #225