@AlekStoyanov 2020-03-15
I could make the nodes bigger. The only downside is that one whole timeline (1 minute of time) won't fit on the screen at once since the gaps left for each second will be bigger. I'll make them bigger and show you screenshots.

About the slowness, it is not due to the filtering, it is due to the drawing of all the elements on the UI.